Adds initial details for MISS runs
Initially, I was planning to just add the research order for the MISS-4 run, but then I realized I have more context to offer!

So, I did a couple of things, and in doing so, realize there are likely going to be changes in the near future that disrupt what
has been done on this day, the day of my dau... no, I don't have a daughter, so.. can't be the day of her wedding! ^_^